+++ include/ws/parser/Parser.hpp
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-#include <ws/parser/Token.hpp>
+#include <ws/parser/token/Token.hpp>
 #include <ws/parser/ParserResult.hpp>
 
 namespace ws::parser {
