+++ include/ws/parser/ast/UnaryOperator.hpp
@@ -2,7 +2,7 @@
 
 #include <memory>
 
-#include <ws/parser/AST.hpp>
+#include <ws/parser/ast/AST.hpp>
 
 namespace ws::parser {
