+++ src/ws/parser/token/TokenType.cpp
@@ -1,17 +1,8 @@
-#pragma once
-
-#include <iostream>
+#include <ws/parser/token/TokenType.hpp>
 
 namespace ws::parser {
 
-enum class TokenType {
-    Parenthesis,
-    Operator,
-    Literal
-
-};
-
-inline std::ostream& operator<<(std::ostream& os, TokenType type) {
+std::ostream& operator<<(std::ostream& os, TokenType type) {
     switch(type) {
         case TokenType::Parenthesis: return os << "parenthesis";
         case TokenType::Operator: return os << "operator";
@@ -20,14 +11,7 @@ inline std::ostream& operator<<(std::ostream& os, TokenType type) {
     }
 }
 
-enum class TokenSubType {
-    Left, Right,
-    Plus, Minus, Multiplication, Division,
-    Float
-
-};
-
-inline std::ostream& operator<<(std::ostream& os, TokenSubType subtype) {
+std::ostream& operator<<(std::ostream& os, TokenSubType subtype) {
     switch(subtype) {
         case TokenSubType::Left: return os << "left";
         case TokenSubType::Right: return os << "right";
